abstracts app fade to own component. cleans up App. renames component and deletes an unused component
